==> 04-missing.Rmd <==
# Missing values

## Data Source 1

```{r, echo=FALSE}
library(dplyr)
library(knitr)
library(UpSetR)
library(naniar)
```


After data cleaning in the last section, we ends up with 1337 rows with at least one missing values, 19422 total missing values, and the percentage of missing values over the entire dataset is around 1.78 percent. 

After that, we would like to know if all variables have the same percentage of missing values. Therefore, we sum up missing values by column as shown below. 

```{r}
data <- read.csv(file='./data/ProcessedData.csv')
data <- data %>% rename(SO2 = TOTAL_SO2, NO3=TOTAL_NO3)
kable(colSums(is.na(data)))
```


```{r}
df <- data %>% select(REGION, SO2, NO3, CA, MG, K, CL)
df_na <- data.frame(percent=round(colSums(is.na(df))/nrow(df)*100))
df_na$keyFeatures <- rownames(df_na)
ggplot(df_na, aes(keyFeatures, percent, fill=percent)) + 
  geom_col()+
  xlab("Features")+
  ylab("Percentage") + 
  ggtitle("Percentage of NAs of Main Features")
```

Notice that each pollutant concentration variables contain almost the same percentage of missing values, about 3 percents. 

```{r}
gg_miss_upset(df)
```

The two plots above are consistent to what we suspected, all pollutant concentrations have almost the same missing value pattern. There are 1259 rows missing all five major pollutants and 16 rows missing four major pollutants. 

## Data Source 2

Let us then look at a cancer dataset we used to study the relationship between air polluants and lung cancer. 
```{r}
load("./data/DiseaseProcessedData.RData")
data <- temp
data[data=="~"] <- NA
kable(colSums(is.na(data)))
```
There are 3077 missing values for three features of the dataset. The following plot provide information on the overall percentage of missing values in each feature. Notice that missing values patterns are uniform across all the entire dataset. 

```{r}
vis_miss(data)
```

If we break the missing values down by a categorical variable called EVENT_TYPE, then we can see that there are are more missing values when the EVENT_TYPE is denoted as "Mortality" than "Incidence". 

```{r}
gg_miss_fct(x=data, fct=EVENT_TYPE)
```
