==> 05-results.Rmd <==
---
title: "Air Pollution over the US"
date: "12/2020"
---


# Results


## Packages
All packages we have used in this section. 
```{r}
library(dplyr)
library(tidyverse)
library(dygraphs)
library(xts)
library(cowplot)
library(GGally)
library(tools)
library(micromap)
library(choroplethr)
```
## Data
Before moving on, let us draw some insights from each pollutant concentration over the past ten years. 
```{r}
data = read.csv(file='./data/processedData.csv')
data <- data %>% na.omit()
data <- data %>% rename(NO3=TOTAL_NO3, SO2=TOTAL_SO2)
data$date = as.Date(data$date_off)
data$year <- format(as.Date(data$date, "%F"), "%Y")
data$month <- format(as.Date(data$date, "%F"), "%m")
data %>% select(SO2, NO3) %>% summary()
```
* Notes:
  + The mean and the median for each pollutant, especially SO2 is far apart. It implies that median might be a better representation to each pollutant distribution. 
  + The largest observation of each pollutant is hugely deviated from mean and median. We need to visualize the distribution of each pollutant more carefully to eliminate any possible outliers. That is what we will do in the next section. 


### Data Distribution
* In order to visualize the distribution of each pollutant, we constructed boxplots. Moreover, since environmental factors vary across the entire nation, we faceted each boxplot by regions (east, middle, west) to see how the concentration of each pollutant varies by regions. For further information about how we split the regions, please refer to previous section called Data Cleaning. 
* Since most data values are lower than 1, we transform all concentration values with log scale to better visualize them. 
```{r}
selected_cols <- c("date", "REGION","STATE", "COUNTY","VALID_HOURS","SO2", "NO3")
new_data <- data %>% select(selected_cols)

tidy_data <- pivot_longer(new_data, 
             cols=c("SO2", "NO3"),
             names_to = "Pollutant",
             values_to = "Concentration",
             values_drop_na = TRUE)

pollutants <- c("SO2", "NO3")

ggplot(tidy_data, aes(x=reorder(Pollutant, -Concentration, median),
                 y = Concentration)) +
  geom_boxplot(fill="#48C9B0", color = "#117864") +
  coord_flip() + 
  scale_y_log10() +
  facet_grid(rows=vars(REGION))+
  xlab("Pollutant") +
  ylab("Concentration") +
  ggtitle("The Distribution of each Air Pollutant faceted by Regions (SO2, NO3)")
```

* Remarks:
  + For both pollutants, the median value is higher in eastern area than the other two regions. In particular, large numbers of high values emerge in that region. Further investigation is needed.
  + There are lots of outliers in each pollutant data. This observation suggests that we should not delete these outliers since they become a commonplace in this analysis. 
  
## How does air pollution spread over the past ten years?

### Time Series of SO2 and NO3 Across the Nation
* Here we plot two dygraphs, one for each pollutant, to visualize the overall distribution in the past ten years. Feel free to play with them and draw some insights. 
```{r}
time_series_obj <- xts(x = data$SO2, order.by = data$date)
dygraph(time_series_obj,
        main = "Sulfur Dioxide Trend in US",
        xlab = "Date",
        ylab = "SO2 (ug/m^3)") %>% 
  dyRangeSelector()
```

```{r}
time_series_obj <- xts(x = data$NO3, order.by = data$date)
dygraph(time_series_obj,
        main = "Nitrate Trend in US",
        xlab = "Date",
        ylab = "NO3 (ug/m^3)") %>% 
  dyRangeSelector()
```

* The overall trends of SO2 and NO3 are similar as they both peak in January 2011 and February 2014. The only different is that the overall trend of SO2 concentration decreases from year to year, while the trend of NO3 remains unchanged. The Environmental Projection Agency and other related organizations have make lots of effort in regulating the emission of SO2 and many common pollutants such as NO2, PM25 since 1970. However, the air pollution problem remains challenging as there are a lot other toxic pollutants such as NO3. Of course, you may complain the time series above are too dense to analyze. In the next section, we will aggregate the concentration of SO2 and NO3 by month to oberve any secular and cyclical trend over time. 

### Monthly Time Series
* Here, we average SO2 and NO3 by months and get the following two monthly plots. 
```{r}
data_month <- data %>% 
  mutate(month_year = paste(year, month, '01', sep='-')) %>% 
  group_by(month_year) %>% 
  summarise(aver_SO2 = mean(SO2), aver_NO3 = mean(NO3)) %>% 
  ungroup()

plot_SO2 <- ggplot(data_month, aes(x=as.Date(month_year, format="%F"), y=aver_SO2)) +
  geom_line(color='#117864') +
  scale_x_date(breaks = seq(as.Date("2010-01-01"), as.Date("2020-08-01"), by="12 months"), date_labels="%Y") +
  xlab("Date") + 
  ylab("SO2 (ug/m^3)") +
  ggtitle("Monthly Average Sulfer Dioxide Concentration")

plot_NO3 <- ggplot(data_month, aes(x=as.Date(month_year, format="%F"), y=aver_NO3)) +
  geom_line(color='#117864') +
  scale_x_date(breaks = seq(as.Date("2010-01-01"), as.Date("2020-08-01"), by="12 months"), date_labels="%Y") +
  xlab("Date") + 
  ylab("NO3 (ug/m^3)") +
  ggtitle("Monthly Average Nitrate Concentration")

plot_grid(plot_SO2, plot_NO3, nrow = 2)
```

* Remarks:
  + Based on the monthly plots above, we may conclude that the overall SO2 concentration decreases year to year. In particular, it decreases significantly in 2020. We suspect part of the reason might be due to Covid-19. In the next section, we will briefly investigate possible effects of Covid-19 on SO2 emission level.  
  + On the other hand, the overall NO3 concentration slightly diminished from 2010 to 2014 and then remain high since then. In addition to common pollutants like SO2, there are actually lots of toxic pollutants similar to NO3 that are challenging our environment. 
  + We can observe cyclic pattern in both plots above. Then the next question we ask is what is causing the cyclical pattern? A rational guess we have is monthly effect. Therefore, we plot monthly sub-series of a time series for each pollutant as follows. 
  
```{r}
data_month$month_year = as.Date(data_month$month_year)
SO2 <- ts(data_month$aver_SO2, start=c(2010, 1), end=c(2020, 8), frequency = 12)
month_SO2 <- monthplot(SO2, main="Sulfer Dioxide Trend, grouped by month")
```

```{r}
NO3 <- ts(data_month$aver_NO3, start=c(2010, 1), end=c(2020, 8), frequency = 12)
month_NO3 <- monthplot(NO3, main="Nitrate Trend, grouped by month")
```

* Based on the two plots above, we observe that monthly (for seasonal) effect causes the cyclical pattern of SO2 since after we facet the plot by month, the overall trends are similar. And the secular trend of SO2 peak in January and February, which makes sense since winter season requires heaters. 
* On the other hand, monthly effect does not entirely cause the cyclical pattern of NO3 as observed above since the overall trends are still different. 

### Did Covid-19 lockdown significantly reduce air pollution in New York State?
* In India and China, researchers found out that air quality improves significantly in urban districts during lockdown. If you want to dive deeper into this research, please refer to [Comparative study on air quality status in India and Chinese cities before and during the COVID-19 lockdown period](https://www.ncbi.nlm.nih.gov/pmc/articles/PMC7375877/){target="_blank"}. By the same interest, we did a brief investigations on the impact of COVID-19 lockdown to the two pollutants, SO2 and NO3. 

* It is unrealistic to investigate the impact of Covid-19 to air pollution across the entire nation since each state responsed differently to the pandemic. Therefore, we decided to choose New York State as a representative. 
The New York State government began with a full lockdown from March 2020 to April 2020 in response to Covid-19 Pandemic. In the following plot, we created a monthly concentration plot by year for each pollutant. Notice in March and April, the SO2 and NO3 concentration level did diminish but not significantly. Therefore, we may conclude that in terms of SO2 and NO3, Covid-19 lockdown did not reduce air pollution significantly. Further investigation is needed on other pollutants. 

```{r}
new_data <- data %>% 
  filter(data$STATE=='NY') %>%
  mutate(month_year = paste(year, month, '01', sep='-')) %>% 
  group_by(month_year) %>% 
  summarise(aver_SO2 = mean(SO2), aver_NO3 = mean(NO3)) %>% 
  ungroup() %>% 
  mutate(
    date = as.Date(month_year),
    xaxis_year = update(date, year = 1),
    year = format(as.Date(date, format="%F"), "%Y")
  )

  ggplot(new_data, aes(xaxis_year, SO2, color = year)) +
    scale_x_date(date_breaks = "1 month", date_labels = "%b") +
    geom_line() +
    xlab("month") +
    ylab("SO2 (ug/m^3)") +
    ggtitle("Monthly SO2 in New York State across each year")
```
```{r}
ggplot(new_data, aes(xaxis_year, NO3, color = year)) +
    scale_x_date(date_breaks = "1 month", date_labels = "%b") +
    geom_line() +
    xlab("month") +
    ylab("NO3 (ug/m^3)") +
    ggtitle("Monthly NO3 in New York State across each year")
```

## How does air pollution spread across states?

### Spatial distribution of SO2 and NO3 over states

* To see the spatial distribution overstates, we draw the linked micromaps over two pollutants, mean SO2 and mean NO3 amount over past 10 years, and show plots of them. 

```{r,fig.width=5, fig.height=6}
#group by state, calculate mean SO2 amount, mean NO3 amount over past 10 years
dataState<- data%>%group_by(STATE)%>%summarise(SO2 = mean(SO2), NO3=mean(NO3))

data("USstates")
statePolys <- create_map_table(USstates, IDcolumn = "ST")

lmplot(stat.data = dataState,
       map.data = statePolys,
       panel.types = c('dot_legend', 'labels', 'dot', 'map'),
       panel.data = list(NA, 'STATE','SO2', NA),
       ord.by = 'SO2', 
       rev.ord = TRUE,
       grouping = 5,
       colors = brewer.pal(5, "Set1"),
       median.row = TRUE,
       map.link = c('STATE', 'ID'),
       map.color2 = "white",
       panel.att = list(
         list(1, point.type=10,
              point.border=TRUE),
          list(2, header = "STATE",
               panel.width = 0.2,
               panel.header.face = "bold",
               panel.header.size = 1,
               graph.grid.color = "grey99",
               align = "left"),
          list(3, header = "Mean SO2 Pollution Amount", 
               panel.header.face = "bold",
               panel.header.size = 1,
               xaxis.title = 'ug/m^3'),
          list(4, header = 'Geographical Location',
               panel.header.face = "bold",
               panel.header.size = 1)
          ))
```

```{r}
dataState<- data%>%group_by(STATE)%>%summarise(SO2 = mean(SO2), NO3=mean(NO3))

#mean so2 by state
dataStateSO2<-dataState%>%select(STATE,SO2)
dataStateSO2<-dataStateSO2%>%rename(region=STATE, value=SO2)
dataStateSO2$region<-tolower(state.name[match(dataStateSO2$region, state.abb)])

state_choropleth(dataStateSO2,title="Mean SO2 Amount Over 10 Years by State", legend="SO2 Amount by State")+ scale_fill_brewer(palette=8)
```

### Spatial distribution of NO3

```{r,fig.width=5, fig.height=6}

data("USstates")
statePolys <- create_map_table(USstates, IDcolumn = "ST")

lmplot(stat.data = dataState,
       map.data = statePolys,
       panel.types = c('dot_legend', 'labels', 'dot', 'map'),
       panel.data = list(NA, 'STATE','NO3', NA),
       ord.by = 'NO3', 
       rev.ord = TRUE,
       grouping = 5,
       colors = brewer.pal(5, "Set1"),
       median.row = TRUE,
       map.link = c('STATE', 'ID'),
       map.color2 = "white",
       panel.att = list(
         list(1, point.type=10,
              point.border=TRUE),
          list(2, header = "STATE",
               panel.width = 0.2,
               panel.header.face = "bold",
               panel.header.size = 1,
               graph.grid.color = "grey99",
               align = "left"),
          list(3, header = "Mean NO3 Pollution Amount", 
               panel.header.face = "bold",
               panel.header.size = 1,
               xaxis.title = 'ug/m^3'),
          list(4, header = 'Geographical Location',
               panel.header.face = "bold",
               panel.header.size = 1)
          ))
```


```{r}
#mean NO3 by state
dataStateNO3<-dataState%>%select(STATE,NO3)
dataStateNO3<-dataStateNO3%>%rename(region=STATE, value=NO3)
dataStateNO3$region<-tolower(state.name[match(dataStateNO3$region, state.abb)])

state_choropleth(dataStateNO3,title="Mean NO3 Amount Over 10 Years by State", legend="NO3 Amount by State")+ scale_fill_brewer(palette=8)
```

* Based on the plots above, we observe that states OH, IN, PA, KY, and VA have the top 5 mean SO2 amount during past 10 years, while IN, IL, OH, NY, KY have top 5 mean NO3 amount during past 10 years.
* According to EPA, the main source of SO2 and NO3 is the burning of fossil fuels and other industrial facilities. Take fossil fuels as example, SO2 and NO3 concentration has a positive correlation with coal and petroleum consumption. According to Primary Energy Consumption Estimates by Source, Ranked by State, of 2018(https://www.eia.gov/state/seds/data.php?incfile=/state/seds/sep_sum/html/rank_use_source.html&sid=US), we noticed that TX, IN, OH have top 3 highest coal consumption, also KY and PA rank at 7th, and 8th in coal consumption. At the same time, for petroleum consumption, states LA, NY, IL, PA, OH rank 3rd, 5th, 6th, 7th, and 8th. These indicate that burning of fossil fuels as coal or petroleum contributes to high SO2 and NO3 amount in states IN, OH, PA, IL, NY and KY. 
* Moreover we found that states with highest SO2 and NO3 amount neighbor each other. This could be affected by the transport of SO2 and NO3 within air shed affected by climate like wind, rain. Emissions of SO2 and NO3 from states OH and IN could travel to the East Coast, contributing to higher amount of SO2 and NO3 pollutants in states such as PA, KY and VA.
* Promoting cleaner energy can make to positive difference to reduce air pollution since coal and petroleum consumption is crucial reason of SO2 and NO3 emission.

### Pollutant distribution over states

```{r}
dataMosaic <- gather(dataState,key = 'pollutant',value = concentration, -STATE)
ggplot(data = dataMosaic, aes(x = reorder(STATE, -concentration), y = concentration, fill = pollutant,order = pollutant)) + 
    geom_bar(stat = "identity") + 
    coord_flip() + 
    ggtitle("Pollutant distribution by states")
```

* This stack bar chart shows the overall distribution of the pollutants, SO2 and NO3. We noticed that states with higher SO2 tend to have higher NO3 emission, but also some states with higher SO2 while have lower NO3 emission. This could be affected by different combination of fossil fuels consuming and factorial emission. Besides energy source such as coal and petroleum, factories and other industrial installations have caused such pollution with metal smelting, petrochemical smog and food processing. For example, heating and smelting metal processes produce emissions of sulfur oxides. 
*  Air pollution can also be controlled by strengthening emission regulation through the installation of filters and scrubbers to clean exhaust fumes from factory processes, and by taking steps to minimize the generation of pollution at the source.

## How does air pollution correlate with Lung and Bronchus cancer?

### Process cancer rate data
Since cancer data is over year 2013-2017, we use the mean of the amount of SO2 and NO3 from 2013-2017 as the indicator.
```{r}
##Combine two datasets
load("./data/DiseaseProcessedData.RData")
data <- read.csv("./data/processedData.csv")
data <- data %>% na.omit()
cancer_data <- temp%>% na.omit()
data$date = as.Date(data$date_off)
data$year <- format(as.Date(data$date, "%F"), "%Y")
data_year <- data %>% filter((year <= 2017) & (year >= 2013))
cancer_data$COUNT <- as.numeric(as.character(cancer_data$COUNT))
cancer_data$POPULATION <- as.numeric(as.character(cancer_data$POPULATION))

incidence_data <- cancer_data%>%
  filter(EVENT_TYPE=="Incidence")
mortality_data <-cancer_data%>%
  filter(EVENT_TYPE=="Mortality")
#Incident count
totalIncidence_count <- incidence_data %>%
  filter(SEX=="Male and Female")%>%
  group_by(STATE)%>%
  summarise(TotalIncidence = sum(COUNT, na.rm = TRUE))
FemaleIncidence_count <- incidence_data %>%
  filter(SEX=='Female')%>%
  group_by(STATE)%>%
  summarise(FemaleIncidence = sum(COUNT, na.rm = TRUE))
MaleIncidence_count <- incidence_data %>%
  filter(SEX=='Male')%>%
  group_by(STATE)%>%
  summarise(MaleIncidence = sum(COUNT, na.rm = TRUE))
#populations 
pop <- incidence_data %>%
  filter(SEX=="Male and Female")%>%
  group_by(STATE)%>%
  summarise(TotalPop = sum(POPULATION, na.rm = TRUE))
Femalepop <- incidence_data %>%
  filter(SEX=="Female")%>%
  group_by(STATE)%>%
  summarise(TotalFPop = sum(POPULATION, na.rm = TRUE))
Malepop <- incidence_data %>%
  filter(SEX=="Male")%>%
  group_by(STATE)%>%
  summarise(TotalMPop = sum(POPULATION, na.rm = TRUE))
#Combine dataframe
PopData <- merge(pop,Femalepop,by ="STATE")%>%
  merge(Malepop,by ="STATE")
CountData <- merge(totalIncidence_count,FemaleIncidence_count,by ="STATE")%>%
  merge(MaleIncidence_count,by ="STATE")
combined <- merge(PopData,CountData,by ="STATE") %>%
  mutate(TotalRate = 100*(TotalIncidence/TotalPop))%>%
  mutate(FemaleRate = 100*(FemaleIncidence/TotalFPop))%>%
  mutate(MaleRate = 100*(MaleIncidence/TotalMPop))
#pollutants
so2 <- data_year %>% 
  group_by(STATE) %>% 
  summarise(TotalSO2 = mean(TOTAL_SO2, na.rm = TRUE))
no3 <- data_year %>% 
  group_by(STATE) %>% 
  summarise(TotalNO3 = mean(TOTAL_NO3, na.rm = TRUE))
region<-data%>%
  select(STATE,REGION)%>%
  distinct(.keep_all = FALSE)
#Cancer and pollutants df
newdata<- merge(x = combined, y = so2, by = "STATE", all.x = TRUE)
newdata<- merge(x = newdata, y = no3, by = "STATE", all.x = TRUE)%>% na.omit()
newdata<- merge(x = newdata, y = region, by = "STATE", all.x = TRUE)%>% na.omit()
newdata$REGION<-as.character(newdata$REGION) 
```

### Correlation between air pollution and Lung and Bronchus cancer rate
```{r}
##Scatter plots 
ggplot(newdata, aes(TotalSO2,TotalRate)) +geom_point(alpha =.5, size =1)  +theme_classic(14) +
  ggtitle("Scatter Plot of Mean SO2 and TotalRate")+
  xlab("meanSO2")
ggplot(newdata, aes(TotalNO3, TotalRate)) +geom_point(alpha =.5, size =1) +theme_classic(14)+
  ggtitle("Scatter Plot of Mean NO3 and TotalRate")+
  xlab("meanNO3")
```

* Remarks:
  + The scatter plots shows that there are no state with high SO2 and low incident rate and no state with low SO2 and high incident rate.
  + The NO3 plot doesn't show a clear pattern. 

```{r}
#cleveland by state
##Cleveland dot plots of all gender cancer rate and SO2, NO3 concentrattion by region

ggplot(newdata, aes(x = TotalRate, y = reorder(STATE, TotalRate),color=REGION)) +
  geom_point(size = 3) +  
  theme_bw() +
  theme(
    panel.grid.major.x = element_blank(),
    panel.grid.minor.x = element_blank(),
    panel.grid.major.y = element_line(colour = "grey60", linetype = "dashed"))+
    ggtitle("Cleveland Dot Plot of All Gender Cancer Region")
ggplot(newdata, aes(x = TotalSO2, y = reorder(STATE, TotalSO2),color=REGION)) +
  geom_point(size = 3) +  
  theme_bw() +
  theme(
    panel.grid.major.x = element_blank(),
    panel.grid.minor.x = element_blank(),
    panel.grid.major.y = element_line(colour = "grey60", linetype = "dashed"))+
    ggtitle("Cleveland Dot Plot of Mean SO2 Concentration")
ggplot(newdata, aes(x = TotalNO3, y = reorder(STATE, TotalNO3),color=REGION)) +
  geom_point(size = 3) +  
  theme_bw() +
  theme(
    panel.grid.major.x = element_blank(),
    panel.grid.minor.x = element_blank(),
    panel.grid.major.y = element_line(colour = "grey60", linetype = "dashed"))+
    ggtitle("Cleveland Dot Plot of Mean NO3 Concentration")
```

* Remarks:
  + The distributions of regions in the above 3 plots have the same pattern: East and West region have higher cancer rate and pollutants' concentrations. Middle region have the lowest cancer rate and pollutants' concentrations. 
  + There is a high chance that Lung and Bronchus cancer rate and pollutants' concentrations are correlated.
  

```{r}
##Correlation matrix with visualizations
ggpairs(newdata[8:12], title="correlogram with ggpairs()")+
  ggtitle("Correlogram of Each Pair of the Indicators")
  
ggcorr(newdata[8:12], method = c("everything", "pearson")) +
  ggtitle("Heatmap of Correlations")
#more related to SO2 than NO3
```

* Remarks:
  + The Corr values of TotalSO2 and TotalRate, TotalSO2 and FemaleRate, TotalSO2 and MaleRate are statistical significantly correlated.
  + The heatmap also shows that the cancer rates are more related to SO2 compared to NO3.


```{r}
## Is there a difference between female cancer rate and male cancer rate?
tidy_data <- pivot_longer(combined, 
             cols=c("FemaleRate", "MaleRate"),
             names_to = "SEX",
             values_to = "RATE",
             values_drop_na = TRUE)
ggplot(tidy_data,aes(RATE,group=1))+
  geom_boxplot()+
  facet_grid(rows=vars(SEX))+
  ggtitle("The Distribution of Cancer Rate Faceted by SEX")
```

* Remarks:
  + The boxplots show that males tend to have a higher Lung and Bronchus cancer incident rate compare to females in general in the US.
  + For female's rate, there is one state as an outlier. 


* Lung and bronchus cancer rate and SO2 concentration are correlated. There is not enough evidence shows that lung and bronchus cancer rate are related to the NO3 concentration. The east and west regions have higher lung and bronchus cancer rates and pollutants' concentrations compare to the middle region of the US. Also males tend to have higher lung and bronchus cancer rate than females.
